api/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers
mysqlclient
pillow
python-decouple
gunicorn
